# Dryft/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Drift API
  description: Backend API for the Drift social/dating application.
  version: 1.0.0
  contact:
    name: Drift Team
servers:
  - url: https://api.drift.app
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /v1/sessions:
    post:
      tags: [Sessions]
      summary: Create a companion session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
...
  /v1/sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Get session details
      operationId: getSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Sessions]
      summary: End a session
      operationId: endSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Session ended
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ended
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        session_code:
          type: string
        host_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, ended]
        participants:
          type: array
          items:
            type: object
            properties:
              user_id:
                type: string
                format: uuid
              device_type:
                type: string
              joined_at:
                type: string
                format: date-time
        created_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
...